requirements.txt
chromadb
sentence-transformers
python-dotenv
openai
pytest
